[CIR][Lowering][NFC] Move helper functions to LoweringHelpers.cpp
This commit moves array initial value lowering relative helper functions
from DirectToLLVM/LowerToLLVM.cpp to LoweringHelpers.cpp.

So ThroughMLIR/LowerCIRToMLIR.cpp can reuse the helper functions to
enable array with initial value lowering in later patch.

This is a refactoring without functional changes.
