+++ clang/lib/CIR/Lowering/CMakeLists.txt
@@ -1,2 +1,42 @@
+set(LLVM_LINK_COMPONENTS
+  Core
+  Support
+  )
+
+get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
+
+add_clang_library(clangCIRLoweringHelpers
+  LoweringHelpers.cpp
+
+  DEPENDS
+  MLIRCIROpsIncGen
+  MLIRCIREnumsGen
+  MLIRCIRASTAttrInterfacesIncGen
+  MLIRCIROpInterfacesIncGen
+  MLIRCIRLoopOpInterfaceIncGen
+  MLIRBuiltinLocationAttributesIncGen
+  MLIRBuiltinTypeInterfacesIncGen
+  MLIRFunctionInterfacesIncGen
+
+  LINK_LIBS
+  clangAST
+  clangBasic
+  clangCodeGen
+  clangLex
+  clangFrontend
+  clangCIR
+  ${dialect_libs}
+  MLIRCIR
+  MLIRAnalysis
+  MLIRBuiltinToLLVMIRTranslation
+  MLIRLLVMToLLVMIRTranslation
+  MLIRIR
+  MLIRParser
+  MLIRSideEffectInterfaces
+  MLIRTransforms
+  MLIRSupport
+  MLIRMemRefDialect
+  )
+
 add_subdirectory(DirectToLLVM)
 add_subdirectory(ThroughMLIR)
